Add constants folder, add FiltersData as constants
