+++ src/mock/index.ts
@@ -1,7 +1,5 @@
 export * from './studentData';
 
-export * from './filterData';
-
 export * from './menuItems';
 
 export * from './notificationsData';
